fix ya make
