+++ ydb/core/fq/libs/checkpointing/ut/checkpoint_coordinator_ut.cpp
@@ -1,4 +1,3 @@
-#include <memory>
 #include <ydb/core/fq/libs/checkpointing/checkpoint_coordinator.h>
 #include <ydb/core/fq/libs/graph_params/proto/graph_params.pb.h>
 #include <ydb/core/testlib/actors/test_runtime.h>
