+++ ydb/library/yql/dq/actors/common/utils.h
@@ -2,7 +2,9 @@
 
 #include <ydb/library/yql/dq/runtime/dq_tasks_runner.h>
 
-namespace NFq {
+namespace NYql::NDq {
+
+NYql::NDqProto::ECheckpointingMode GetTaskCheckpointingMode(const NYql::NDq::TDqTaskSettings& task);
 
 bool IsIngress(const NYql::NDq::TDqTaskSettings& task);
 
@@ -10,4 +12,4 @@ bool IsEgress(const NYql::NDq::TDqTaskSettings& task);
 
 bool HasState(const NYql::NDq::TDqTaskSettings& task);
 
-} // namespace NFq
+} // namespace NYql::NDq
