+++ ydb/library/yql/dq/actors/common/ya.make
@@ -2,12 +2,14 @@ LIBRARY()
 
 SRCS(
     retry_queue.cpp
+    utils.cpp
 )
 
 PEERDIR(
     ydb/library/actors/core
     ydb/library/yql/dq/actors/protos
     yql/essentials/public/issue
+    ydb/library/yql/dq/runtime
 )
 
 YQL_LAST_ABI_VERSION()
